✨ 아티스트 관리 페이지 컴포넌트 리팩토링 및 새로운 구조로 변경
